feat: add in roles module and functionality
