+++ src/interaction/events/guild/guildMemberAdd.ts
@@ -9,7 +9,7 @@ import { LogAction } from '@prisma/client';
 export default async function guildMemberAdd(auxdibot: Auxdibot, member: GuildMember) {
    if (!member) return;
    const server = await findOrCreateServer(auxdibot, member.guild.id);
-   if (server.join_leave_channel) {
+   if (server.join_leave_channel && server.disabled_modules.indexOf('Greetings') == -1) {
       member.guild.channels.fetch(server.join_leave_channel).then(async (channel) => {
          if (channel && channel.isTextBased()) {
             if (server.join_embed || server.join_text) {
@@ -36,7 +36,7 @@ export default async function guildMemberAdd(auxdibot: Auxdibot, member: GuildMe
          }
       });
    }
-   if (server.join_dm_embed || server.join_dm_text) {
+   if ((server.join_dm_embed || server.join_dm_text) && server.disabled_modules.indexOf('Greetings') == -1) {
       await member
          .send({
             content: `${server.join_dm_text || ''}`,
@@ -52,9 +52,13 @@ export default async function guildMemberAdd(auxdibot: Auxdibot, member: GuildMe
          })
          .catch(() => undefined);
    }
-   if (server.join_roles.length > 0)
-      server.join_roles.forEach((role: string) => member.roles.add(role).catch(() => undefined));
-   memberJoin(auxdibot, member.guild.id, member);
+
+   if (server.disabled_modules.indexOf('Roles') == -1) {
+      if (server.join_roles.length > 0)
+         server.join_roles.forEach((role: string) => member.roles.add(role).catch(() => undefined));
+      memberJoin(auxdibot, member.guild.id, member);
+   }
+
    await handleLog(
       auxdibot,
       member.guild,
