+++ src/interaction/subcommands/greetings/join/joinPreview.ts
@@ -6,27 +6,27 @@ import { AuxdibotSubcommand } from '@/interfaces/commands/AuxdibotSubcommand';
 import handleError from '@/util/handleError';
 import parsePlaceholders from '@/util/parsePlaceholder';
 
-export const leavePreview = <AuxdibotSubcommand>{
+export const joinPreview = <AuxdibotSubcommand>{
    name: 'preview',
    info: {
-      module: Modules['Settings'],
-      description: 'Preview the leave message.',
-      usageExample: '/leave preview',
-      permission: 'settings.leave.preview',
+      module: Modules['Greetings'],
+      description: 'Preview the join message.',
+      usageExample: '/join preview',
+      permission: 'greetings.join.preview',
    },
    async execute(auxdibot: Auxdibot, interaction: AuxdibotCommandInteraction<GuildAuxdibotCommandData>) {
       if (!interaction.data) return;
-      const settings = interaction.data.guildData;
+      const server = interaction.data.guildData;
       try {
          return await interaction.reply({
-            content: `**EMBED PREVIEW**\r\n${settings.leave_text || ''}`,
-            ...(Object.entries(settings.leave_embed || {}).length != 0
+            content: `**EMBED PREVIEW**\r\n${server.join_text || ''}`,
+            ...(Object.entries(server.join_embed || {}).length != 0
                ? {
                     embeds: [
                        JSON.parse(
                           await parsePlaceholders(
                              auxdibot,
-                             JSON.stringify(settings.leave_embed),
+                             JSON.stringify(server.join_embed),
                              interaction.data.guild,
                              interaction.data.member,
                           ),
@@ -38,7 +38,7 @@ export const leavePreview = <AuxdibotSubcommand>{
       } catch (x) {
          return await handleError(
             auxdibot,
-            'INVALID_LEAVE_EMBED',
+            'INVALID_JOIN_EMBED',
             'This is an invalid embed! This will not be able to send when a user joins the server. Change either the Embed or Text to fix this error.',
             interaction,
          );
