+++ src/interaction/subcommands/greetings/join_dm/joinDMEmbedJSON.ts
@@ -3,30 +3,30 @@ import { Auxdibot } from '@/interfaces/Auxdibot';
 import { GuildAuxdibotCommandData } from '@/interfaces/commands/AuxdibotCommandData';
 import AuxdibotCommandInteraction from '@/interfaces/commands/AuxdibotCommandInteraction';
 import { AuxdibotSubcommand } from '@/interfaces/commands/AuxdibotSubcommand';
-import setJoinEmbed from '@/modules/features/greetings/setJoinEmbed';
+import setJoinDMEmbed from '@/modules/features/greetings/setJoinDMEmbed';
 import handleError from '@/util/handleError';
 import parsePlaceholders from '@/util/parsePlaceholder';
 import { EmbedBuilder } from '@discordjs/builders';
 import { APIEmbed } from '@prisma/client';
 
-export const joinEmbedJSON = <AuxdibotSubcommand>{
+export const joinDMEmbedJSON = <AuxdibotSubcommand>{
    name: 'embed_json',
    info: {
-      module: Modules['Settings'],
+      module: Modules['Greetings'],
       description:
-         'Add an embed to the join message using custom JSON. (Placeholders are supported. Do /help placeholders for a list of placeholders.)',
-      usageExample: '/join embed_json (json)',
-      permission: 'settings.join.embed_json',
+         'Add an embed to the join DM message using custom JSON. (Placeholders are supported. Do /help placeholders for a list of placeholders.)',
+      usageExample: '/join_dm embed_json (json)',
+      permission: 'greetings.join_dm.embed_json',
    },
    async execute(auxdibot: Auxdibot, interaction: AuxdibotCommandInteraction<GuildAuxdibotCommandData>) {
       if (!interaction.data) return;
       const json = interaction.options.getString('json', true);
       const server = interaction.data.guildData;
       try {
-         const jsonEmbed = JSON.parse(json) as APIEmbed;
+         const jsonEmbed = JSON.parse(json) satisfies APIEmbed;
          if (interaction.channel && interaction.channel.isTextBased())
             await interaction.channel.send({
-               content: "Here's a preview of the new join embed!",
+               content: "Here's a preview of the new join DM embed!",
                ...(Object.entries(json || {}).length != 0
                   ? {
                        embeds: [
@@ -37,10 +37,10 @@ export const joinEmbedJSON = <AuxdibotSubcommand>{
                     }
                   : {}),
             });
-         await setJoinEmbed(auxdibot, server.serverID, jsonEmbed);
+         await setJoinDMEmbed(auxdibot, server.serverID, jsonEmbed);
          const embed = new EmbedBuilder().setColor(auxdibot.colors.accept).toJSON();
          embed.title = 'Success!';
-         embed.description = `Set the join embed.`;
+         embed.description = `Set the join DM embed.`;
          return await interaction.reply({ embeds: [embed] });
       } catch (x) {
          return await handleError(
