+++ src/interaction/subcommands/greetings/join_dm/joinDMPreview.ts
@@ -6,27 +6,27 @@ import { AuxdibotSubcommand } from '@/interfaces/commands/AuxdibotSubcommand';
 import handleError from '@/util/handleError';
 import parsePlaceholders from '@/util/parsePlaceholder';
 
-export const joinPreview = <AuxdibotSubcommand>{
+export const joinDMPreview = <AuxdibotSubcommand>{
    name: 'preview',
    info: {
-      module: Modules['Settings'],
-      description: 'Preview the join message.',
-      usageExample: '/join preview',
-      permission: 'settings.join.preview',
+      module: Modules['Greetings'],
+      description: 'Preview the join DM message.',
+      usageExample: '/join_dm preview',
+      permission: 'greetings.join_dm.preview',
    },
    async execute(auxdibot: Auxdibot, interaction: AuxdibotCommandInteraction<GuildAuxdibotCommandData>) {
       if (!interaction.data) return;
-      const server = interaction.data.guildData;
+      const settings = interaction.data.guildData;
       try {
          return await interaction.reply({
-            content: `**EMBED PREVIEW**\r\n${server.join_text || ''}`,
-            ...(Object.entries(server.join_embed || {}).length != 0
+            content: `**EMBED PREVIEW**\r\n${settings.join_dm_text || ''}`,
+            ...(Object.entries(settings.join_dm_embed || {}).length != 0
                ? {
                     embeds: [
                        JSON.parse(
                           await parsePlaceholders(
                              auxdibot,
-                             JSON.stringify(server.join_embed),
+                             JSON.stringify(settings.join_dm_embed),
                              interaction.data.guild,
                              interaction.data.member,
                           ),
@@ -38,7 +38,7 @@ export const joinPreview = <AuxdibotSubcommand>{
       } catch (x) {
          return await handleError(
             auxdibot,
-            'INVALID_JOIN_EMBED',
+            'INVALID_JOIN_DM_EMBED',
             'This is an invalid embed! This will not be able to send when a user joins the server. Change either the Embed or Text to fix this error.',
             interaction,
          );
