+++ src/interaction/subcommands/greetings/leave/leaveMessage.ts
@@ -14,12 +14,12 @@ import setLeaveEmbed from '@/modules/features/greetings/setLeaveEmbed';
 export const leaveMessage = <AuxdibotSubcommand>{
    name: 'message',
    info: {
-      module: Modules['Settings'],
+      module: Modules['Greetings'],
       description:
          'Set the leave message. (Placeholders are supported. Do /help placeholders for a list of placeholders.)',
       usageExample:
          '/leave message [content] [color] [title] [title url] [author] [author icon url] [author url] [description] [fields (split title and description with "|d|", and seperate fields with "|s|")] [footer] [footer icon url] [image url] [thumbnail url]',
-      permission: 'settings.leave.message',
+      permission: 'greetings.leave.message',
    },
    async execute(auxdibot: Auxdibot, interaction: AuxdibotCommandInteraction<GuildAuxdibotCommandData>) {
       if (!interaction.data) return;
