+++ src/interaction/subcommands/greetings/leave/leavePreview.ts
@@ -6,27 +6,27 @@ import { AuxdibotSubcommand } from '@/interfaces/commands/AuxdibotSubcommand';
 import handleError from '@/util/handleError';
 import parsePlaceholders from '@/util/parsePlaceholder';
 
-export const joinDMPreview = <AuxdibotSubcommand>{
+export const leavePreview = <AuxdibotSubcommand>{
    name: 'preview',
    info: {
-      module: Modules['Settings'],
-      description: 'Preview the join DM message.',
-      usageExample: '/join_dm preview',
-      permission: 'settings.joindm.preview',
+      module: Modules['Greetings'],
+      description: 'Preview the leave message.',
+      usageExample: '/leave preview',
+      permission: 'greetings.leave.preview',
    },
    async execute(auxdibot: Auxdibot, interaction: AuxdibotCommandInteraction<GuildAuxdibotCommandData>) {
       if (!interaction.data) return;
       const settings = interaction.data.guildData;
       try {
          return await interaction.reply({
-            content: `**EMBED PREVIEW**\r\n${settings.join_dm_text || ''}`,
-            ...(Object.entries(settings.join_dm_embed || {}).length != 0
+            content: `**EMBED PREVIEW**\r\n${settings.leave_text || ''}`,
+            ...(Object.entries(settings.leave_embed || {}).length != 0
                ? {
                     embeds: [
                        JSON.parse(
                           await parsePlaceholders(
                              auxdibot,
-                             JSON.stringify(settings.join_dm_embed),
+                             JSON.stringify(settings.leave_embed),
                              interaction.data.guild,
                              interaction.data.member,
                           ),
@@ -38,7 +38,7 @@ export const joinDMPreview = <AuxdibotSubcommand>{
       } catch (x) {
          return await handleError(
             auxdibot,
-            'INVALID_JOIN_DM_EMBED',
+            'INVALID_LEAVE_EMBED',
             'This is an invalid embed! This will not be able to send when a user joins the server. Change either the Embed or Text to fix this error.',
             interaction,
          );
