+++ src/interaction/subcommands/roles/join_roles/joinRoleAdd.ts
@@ -11,34 +11,33 @@ import { EmbedBuilder } from '@discordjs/builders';
 import { LogAction } from '@prisma/client';
 import { PermissionsBitField } from 'discord.js';
 
-export const stickyRoleAdd = <AuxdibotSubcommand>{
+export const joinRoleAdd = <AuxdibotSubcommand>{
    name: 'add',
    info: {
-      module: Modules['Settings'],
-      description: 'Add a role to be kept when a member rejoins the server.',
-      usageExample: '/sticky_roles add (role)',
-      permission: 'settings.sticky_roles.add',
+      module: Modules['Roles'],
+      description: 'Add a role to be assigned when a member joins the server.',
+      usageExample: '/join_roles add (role)',
+      permission: 'roles.join_roles.add',
    },
    async execute(auxdibot: Auxdibot, interaction: AuxdibotCommandInteraction<GuildAuxdibotCommandData>) {
       if (!interaction.data || !interaction.memberPermissions) return;
-      const role = interaction.options.getRole('role', true);
       const server = interaction.data.guildData;
+      const role = interaction.options.getRole('role', true);
       if (role.id == interaction.data.guild.roles.everyone.id) {
-         return await handleError(auxdibot, 'STICKY_ROLE_EVERYONE', 'This is the everyone role, silly!', interaction);
+         return await handleError(auxdibot, 'JOIN_ROLE_EVERYONE', 'This is the everyone role, silly!', interaction);
       }
-      if (server.sticky_roles.find((val: string) => role != null && val == role.id)) {
+      if (server.join_roles.find((val: string) => role != null && val == role.id)) {
          return await handleError(
             auxdibot,
-            'STICKY_ROLE_EXISTS',
-            'This role is already added as a sticky role!',
+            'JOIN_ROLE_EXISTS',
+            'This role is already added as a join role!',
             interaction,
          );
       }
       if (
-         role &&
          interaction.data.member.id != interaction.data.guild.ownerId &&
          !interaction.memberPermissions.has(PermissionsBitField.Flags.Administrator) &&
-         interaction.data.guild.roles.comparePositions(role.id, interaction.data.member.roles.highest) <= 0
+         role.position >= interaction.data.member.roles.highest.position
       ) {
          return await handleError(
             auxdibot,
@@ -48,9 +47,8 @@ export const stickyRoleAdd = <AuxdibotSubcommand>{
          );
       }
       if (
-         role &&
          interaction.data.guild.members.me &&
-         interaction.data.guild.roles.comparePositions(role.id, interaction.data.guild.members.me.roles.highest) >= 0
+         role.position >= interaction.data.guild.members.me.roles.highest.position
       ) {
          return await handleError(
             auxdibot,
@@ -59,25 +57,20 @@ export const stickyRoleAdd = <AuxdibotSubcommand>{
             interaction,
          );
       }
-      if (!testLimit(server.sticky_roles, Limits.STICKY_ROLE_DEFAULT_LIMIT)) {
-         return await handleError(
-            auxdibot,
-            'STICKY_ROLES_LIMIT_EXCEEDED',
-            'You have too many sticky roles!',
-            interaction,
-         );
+      if (!testLimit(server.join_roles, Limits.JOIN_ROLE_DEFAULT_LIMIT)) {
+         return await handleError(auxdibot, 'JOIN_ROLES_LIMIT_EXCEEDED', 'You have too many join roles!', interaction);
       }
       await auxdibot.database.servers.update({
          where: { serverID: server.serverID },
-         data: { sticky_roles: { push: role.id } },
+         data: { join_roles: { push: role.id } },
       });
       const successEmbed = new EmbedBuilder().setColor(auxdibot.colors.accept).toJSON();
-      successEmbed.title = '📝 Added Sticky Role';
-      successEmbed.description = `Added <@&${role.id}> to the sticky roles.`;
+      successEmbed.title = '👋 Added Join Role';
+      successEmbed.description = `Added <@&${role.id}> to the join roles.`;
       await handleLog(auxdibot, interaction.data.guild, {
          userID: interaction.data.member.id,
-         description: `Added ${role.name} to sticky roles.`,
-         type: LogAction.STICKY_ROLE_ADDED,
+         description: `Added (Role ID: ${role.id}) to the join roles.`,
+         type: LogAction.JOIN_ROLE_ADDED,
          date_unix: Date.now(),
       });
       return await interaction.reply({ embeds: [successEmbed] });
