+++ src/interaction/subcommands/roles/join_roles/joinRoleRemove.ts
@@ -12,11 +12,11 @@ import { PermissionsBitField } from 'discord.js';
 export const joinRoleRemove = <AuxdibotSubcommand>{
    name: 'remove',
    info: {
-      module: Modules['Settings'],
+      module: Modules['Roles'],
       description:
          "Remove a role that is assigned when a member joins the server. If you've deleted the role, use the index parameter, which is the placement of the item on /join_roles list.",
       usageExample: '/join_roles remove [role] [index]',
-      permission: 'settings.join_roles.remove',
+      permission: 'roles.join_roles.remove',
    },
    async execute(auxdibot: Auxdibot, interaction: AuxdibotCommandInteraction<GuildAuxdibotCommandData>) {
       if (!interaction.data || !interaction.memberPermissions) return;
