+++ src/interaction/subcommands/roles/reactionRoles/reactionRolesAddJSON.ts
